==> notebooks/12_load_features_hopsworks.ipynb <==
from dotenv import load_dotenv
import os
load_dotenv() 

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

from datetime import datetime
import pandas as pd
from src.data_utils import load_and_process_taxi_data

from_year = 2023
# to_year = datetime.now().year
to_year = 2024
print(f"Download raw data from {from_year} to {to_year}")

rides = pd.DataFrame()
chunks = []
for year in range(from_year, to_year+1):

    rides_one_year = load_and_process_taxi_data(year)

    chunks.append(rides_one_year)
    break

# Concatenate all chunks at the end
rides = pd.concat(chunks, ignore_index=True)
print("Data loading complete.")

rides

rides.shape

from src.data_utils import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)

ts_data.shape

ts_data.info()

ts_data

import os
import hopsworks

# Ensure API Key and Project Name are set correctly
os.environ["HOPSWORKS_API_KEY"] = '02pLg96cdTe6EnR5.uQmX3w8fPogt5m1OVMhxJxgVXCMwyEznBF0HZZkPELUWQxChbkqpkJFM6mp1j63p'  # Replace with your actual API key
os.environ["HOPSWORKS_PROJECT_NAME"] = 'ny_taxi_manognat'

# Force execution engine to Python
os.environ["HOPSWORKS_ENGINE"] = "python"

# Reconnect to Hopsworks
project = hopsworks.login(
    api_key_value= '02pLg96cdTe6EnR5.uQmX3w8fPogt5m1OVMhxJxgVXCMwyEznBF0HZZkPELUWQxChbkqpkJFM6mp1j63p',
    project= 'ny_taxi_manognat')

print(f"Successfully reconnected to Hopsworks project: {project.name}")



# import logging
# import hopsworks

# logging.basicConfig(level=logging.DEBUG)

# project = hopsworks.login(
#     api_key_value="diQgze5Ia3gKj0lG.OhMVWQZt2EpYdxj03yCMIFz7ozaPFcteGIYUyQaQRdKRFwrS9TvXERo57gVdxSWG",
#     project="ny_taxi_manognat",
#     engine="python"
# )



feature_store = project.get_feature_store()

FEATURE_GROUP_NAME = "time_series_hourly_feature_group"
FEATURE_GROUP_VERSION = 1

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description = "Time-series data at hourly frequency",
    primary_key = ["pickup_location_id", "pickup_hour"],
    event_time = "pickup_hour"
)

feature_group.insert(ts_data, write_options={"wait_for_job": False})

df_memory_mb = rides.memory_usage(deep=True).sum() / (1024 * 1024)  
print(f"DataFrame size: {df_memory_mb:.2f} MB")

ts_data.info()
